CASSANDRA-9318 Fixes and improvements:
* Back-pressure is now applied before sending any mutations in StorageProxy.
* Back-pressure is now reset if no mutations have been sent in the time window: this is to avoid to keep rate limiting even after "quiet" periods.
* The window precision is now computed in relation to its size.

+++ src/java/org/apache/cassandra/utils/SlidingTimeRate.java
@@ -46,7 +46,7 @@ public class SlidingTimeRate
      * The precision defines how accurate the rate computation is, as it will be computed over window size +/-
      * precision.
      */
-    public SlidingTimeRate(TimeSource timeSource, long size, int precision, TimeUnit unit)
+    public SlidingTimeRate(TimeSource timeSource, long size, long precision, TimeUnit unit)
     {
         Preconditions.checkArgument(size > precision, "Size should be greater than precision.");
         Preconditions.checkArgument(TimeUnit.MILLISECONDS.convert(precision, unit) >= 1, "Precision must be greater than or equal to 1 millisecond.");
